+++ rpg-api/src/main/java/com/github/_255_ping/rpg/api/items/BuiltinItemType.java
@@ -4,6 +4,7 @@ public enum BuiltinItemType implements ItemType {
     SWORD,
     WAND,
     BOW,
+    CROSSBOW,
     ARMOR,
     MATERIAL,
     QUEST,